fix(FEC-13594): playlist with youtube blocks interaction with player (#750)
### Description of the Changes

- bug fix.

**the issue:**
when loading a playlist with youtube entry and media that has dualscreen, and playing the dualscreen media- the dualscreen player is not clickable. in addition, the interaction with the player gui area is also "blocked".

**root cause:**
we are adding `z-index = 1` to the engine element in case it is a youtube engine. when switching to the dual media (or just not a YT entry), it is blocking the interaction with the player in the gui area (play/pause), which includes blocking interaction with the secondary player (aka the dualscreen).

**solution:**
remove the value of `z-index` in case the engine is not youtube.

Solves FEC-13594